sdk/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-evaluation-sdk"
version = "0.1.0"
description = "Production-ready evaluation infrastructure for AI agents"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Ahmed Yasser Eita", email = "ahmed@example.com"}
]
keywords = ["ai", "agents", "evaluation", "gcp", "adk", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-cloud-logging>=3.10.0",
    "google-cloud-trace>=1.13.0",
    "google-cloud-monitoring>=2.21.0",
    "google-cloud-bigquery>=3.21.0",
    "google-cloud-aiplatform>=1.60.0",
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-PyYAML>=6.0.0",
    "types-protobuf>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/agent-evaluation-agent"
Documentation = "https://github.com/yourusername/agent-evaluation-agent/docs"
Repository = "https://github.com/yourusername/agent-evaluation-agent"
Issues = "https://github.com/yourusername/agent-evaluation-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_evaluation_sdk*"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Allow some untyped functions
ignore_missing_imports = true  # Ignore missing stubs for Google Cloud libraries

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=agent_evaluation_sdk --cov-report=html --cov-report=term"

